Functions to parse instructions from files
In the emulator module, implemented `get_instructions` and `sanitize`
which read a file and sanitize it, respectively. This puts the
instructions in the file into a list of integers which can easily be
executed by the emulator.

TODO: `sanitize` function needs a cleanup and optimise
TODO: add other input types such as CLI and stream