Implemented ability to add breakpoints
Within the instruction sequence placing a `|` will signal a breakpoints.
When a breakpoint is being executed, the microprocessor will halt and
the values for IP, IS, R0 and R1 will be shown in the terminal. Then the
user will be prompted to press enter when they want to continue the
execution of the program.