Tests for and fixed loading instructions
Starting writing tests for the emulator's `load_instructions` function
before realising that it had a bug where by the locations of the
breakpoints would be stored into the array as a 0. This is detrementat
as 0 is the halt instruction and so the execution of the sequence
would've stopped.

OK: 19 tests